create proxy mod, move Item proxy there
